+++ pages/posts/index.vue
@@ -1,36 +1,36 @@
 <template>
   <div class="container mx-auto px-2 md:px-12 lg:px-16 xl:px-20">
-    <ul class="flex text-xs border border-black">
+    <ul class="flex text-xs border border-black overflow-x-scroll">
       <li
-        class="w-32 hover:bg-black hover:text-white"
-        :class="{ 'bg-black text-white': category == '' }"
+        class=" hover:bg-black hover:text-white"
+        :class="{ 'bg-black text-white': '' == categoryCode }"
       >
         <button
           aria-label="posts_all"
           type="button"
-          @click="(category = ''), $fetch(), (page = 0)"
-          class="w-full h-10 font-bold uppercase focus:outline-none"
+          @click="retrieve(0, '')"
+          class="w-32 h-10 font-bold uppercase focus:outline-none"
         >
           All
         </button>
       </li>
       <li
-        class="w-32 hover:bg-black hover:text-white"
-        :class="{ 'bg-black text-white': category == cg.alias }"
-        v-for="(cg, index) in categories"
-        :key="index"
+        class="hover:bg-black hover:text-white"
+        :class="{ 'bg-black text-white': cg.code == categoryCode }"
+        v-for="cg in categories"
+        :key="cg.code"
       >
         <button
           :aria-label="'posts_' + cg.alias"
           type="button"
-          @click="(category = cg.alias), $fetch(), (page = 0)"
-          class="w-full h-10 font-bold uppercase focus:outline-none"
+          @click="retrieve(0, cg.code)"
+          class="w-32 h-10 font-bold uppercase focus:outline-none"
           v-text="cg.alias"
         ></button>
       </li>
     </ul>
     <PostsList :datas="datas" />
-    <Pagation :page="page" :total="total" @retrieve="retrieve" />
+    <Pagation :page="page" :size="size" :total="total" @retrieve="retrieve" />
   </div>
 </template>
 
@@ -53,32 +53,34 @@ export default defineComponent({
 
   setup() {
     const route = useRoute();
-    // 获取路由参数
-    const category = computed(() => route.value.query.category);
-    // 匹配类目code
-    const code = computed(() => {
+    const category = computed(() => {
       let data: any = categories.value.filter(
-        (item: any) => category.value == item.alias
+        (item: any) => route.value.query.category == item.alias
       );
       return data.code ? data.code : "";
     });
+
+    // cotegory code
+    const categoryCode = ref("");
 
     const categories = ref([]);
     const datas = ref([]);
 
     // 分页参数
     const page = ref(0);
+    const size = ref(12);
     const total = ref(0);
 
     const { $axios } = useContext();
 
-    const { fetch } = useFetch(async () => {
+    useFetch(async () => {
       [categories.value, datas.value, total.value] = await Promise.all([
         $axios.$get(SERVER_URL.category),
         $axios.$get(
           SERVER_URL.posts.concat(
             "?page=" + page.value,
-            "&size=12&category=" + code.value
+            "&size=" + size.value,
+            "&category=" + category.value
           )
         ),
         $axios.$get(SERVER_URL.posts.concat("/count")),
@@ -103,17 +105,29 @@ export default defineComponent({
       ],
     }));
 
-    const retrieve = (num: number) => {
+    const retrieve = (num: number, code: string) => {
       page.value = num;
-      fetch();
+      if (code) {
+        categoryCode.value = code;
+      }
+      $axios
+        .get(
+          SERVER_URL.posts.concat(
+            "?page=" + page.value,
+            "&size=" + size.value,
+            "&category=" + categoryCode.value
+          )
+        )
+        .then((res) => (datas.value = res.data));
     };
 
     return {
-      category,
       categories,
       datas,
+      categoryCode,
 
       page,
+      size,
       total,
 
       retrieve,